Add DateBox component as part of refactoring SubstituteManagement.
