+++ client/src/view/tutorial-substitutes/SubstituteManagement.tsx
@@ -9,46 +9,27 @@ import {
   Tooltip,
   Typography,
 } from '@material-ui/core';
-import { SelectInputProps } from '@material-ui/core/Select/SelectInput';
 import { createStyles, fade, makeStyles } from '@material-ui/core/styles';
 import { plainToClass } from 'class-transformer';
 import _ from 'lodash';
 import { DateTime } from 'luxon';
-import {
-  AccountSearch as SearchIcon,
-  ChevronRight as RightArrow,
-  Close as RemoveIcon,
-} from 'mdi-material-ui';
-import React, { useCallback, useEffect, useState } from 'react';
+import { AccountSearch as SearchIcon, Close as RemoveIcon } from 'mdi-material-ui';
+import React, { useCallback, useState } from 'react';
 import { NamedElement } from 'shared/model/Common';
 import { ITutorial } from 'shared/model/Tutorial';
 import { getNameOfEntity, sortByName } from 'shared/util/helpers';
 import BackButton from '../../components/BackButton';
-import CustomSelect from '../../components/CustomSelect';
 import DateOrIntervalText from '../../components/DateOrIntervalText';
 import OutlinedBox from '../../components/OutlinedBox';
 import { Tutorial } from '../../model/Tutorial';
 import { ROUTES } from '../../routes/Routing.routes';
-import { compareDateTimes } from '../../util/helperFunctions';
+import DateBox from './components/DateBox';
 
 const useStyles = makeStyles((theme) =>
   createStyles({
-    scrollableBox: {
-      overflowY: 'auto',
-      ...theme.mixins.scrollbar(8),
-    },
     backButton: {
       height: 'fit-content',
       marginRight: theme.spacing(1),
-    },
-    dateButton: {
-      marginTop: theme.spacing(1.5),
-      '&:first-of-type': {
-        marginTop: 0,
-      },
-    },
-    dateButtonIcon: {
-      marginLeft: 'auto',
     },
     selectedSubstitute: {
       display: 'grid',
@@ -75,13 +56,6 @@ const useStyles = makeStyles((theme) =>
   })
 );
 
-enum FilterOption {
-  ONLY_FUTURE_DATES = 'Nur zukünftige Termine',
-  ALL_DATES = 'Alle Termine',
-  WITHOUT_SUBSTITUTE = 'Ohne Vertretung',
-  WITH_SUBSTITUTE = 'Mit Vertretung',
-}
-
 interface Props {
   tutorial: Tutorial;
   students: NamedElement[];
@@ -102,45 +76,15 @@ function filterStudents(students: NamedElement[], filterText: string): NamedElem
     .sort(sortByName);
 }
 
-function filterDates(tutorial: Tutorial, option: FilterOption): DateTime[] {
-  return tutorial.dates
-    .filter((date) => {
-      switch (option) {
-        case FilterOption.ONLY_FUTURE_DATES:
-          const diff = date.startOf('days').diffNow('days').days;
-          return Math.ceil(diff) >= 0;
-        case FilterOption.ALL_DATES:
-          return true;
-        case FilterOption.WITH_SUBSTITUTE:
-          return tutorial.getSubstitute(date) !== undefined;
-        case FilterOption.WITHOUT_SUBSTITUTE:
-          return tutorial.getSubstitute(date) === undefined;
-        default:
-          return true;
-      }
-    })
-    .sort(compareDateTimes);
-}
-
-function TutorialSubstituteManagementContent({ tutorial, students }: Props): JSX.Element {
+function SubstituteManagementContent({ tutorial, students }: Props): JSX.Element {
   const classes = useStyles();
 
-  const [filterOption, setFilterOption] = useState<FilterOption>(
-    () => FilterOption.ONLY_FUTURE_DATES
-  );
   const [filterText, setFilterText] = useState('');
 
-  const [datesToShow, setDatesToShow] = useState<DateTime[]>(() =>
-    filterDates(tutorial, filterOption)
-  );
   const [studentsToShow, setStudentsToShow] = useState<NamedElement[]>(() =>
     filterStudents(students, filterText)
   );
   const [selectedDate, setSelectedDate] = useState<DateTime>();
-
-  useEffect(() => {
-    setDatesToShow(filterDates(tutorial, filterOption));
-  }, [tutorial, filterOption]);
 
   const debouncedHandleChange = useCallback(
     _.debounce(
@@ -149,22 +93,6 @@ function TutorialSubstituteManagementContent({ tutorial, students }: Props): JSX
     ),
     []
   );
-
-  const handleChange: SelectInputProps['onChange'] = (e) => {
-    if (typeof e.target.value !== 'string') {
-      return;
-    }
-
-    const selectedOption: FilterOption | undefined = Object.values(FilterOption).find(
-      (op) => op === e.target.value
-    );
-
-    if (!selectedOption) {
-      throw new Error('Selected filter option is not a valid one.');
-    }
-
-    setFilterOption(selectedOption);
-  };
 
   const handleTextChange: InputProps['onChange'] = (e) => {
     setFilterText(e.target.value);
@@ -189,60 +117,7 @@ function TutorialSubstituteManagementContent({ tutorial, students }: Props): JSX
         </Typography>
       </Box>
 
-      <Box display='flex' flexDirection='column' className={classes.scrollableBox}>
-        <Typography variant='h6' style={{ marginBottom: 8 * 1.5 }}>
-          Datum auswählen
-        </Typography>
-
-        <CustomSelect
-          label='Filtern'
-          items={Object.values(FilterOption)}
-          itemToValue={(item) => item}
-          itemToString={(item) => item}
-          value={filterOption}
-          onChange={handleChange}
-        />
-
-        <OutlinedBox
-          flex={1}
-          display='flex'
-          flexDirection='column'
-          marginTop={1.5}
-          padding={1}
-          className={classes.scrollableBox}
-        >
-          {datesToShow.map((date) => (
-            <Button
-              variant='outlined'
-              color={date === selectedDate ? 'primary' : 'default'}
-              key={date.toISODate()}
-              className={classes.dateButton}
-              classes={{ endIcon: classes.dateButtonIcon }}
-              endIcon={<RightArrow />}
-              onClick={() => setSelectedDate(date)}
-            >
-              <Box
-                display='flex'
-                flexDirection='column'
-                textAlign='left'
-                style={{ textTransform: 'none' }}
-              >
-                <DateOrIntervalText date={date} />
-                {
-                  <Typography variant='caption'>
-                    {!!tutorial.getSubstitute(date)
-                      ? `Vertretung: ${getNameOfEntity(tutorial.getSubstitute(date)!)}`
-                      : 'Keine Vertretung'}
-                  </Typography>
-                }
-              </Box>
-            </Button>
-          ))}
-          {datesToShow.length === 0 && (
-            <Typography align='center'>Keine entsprechenden Termine gefunden.</Typography>
-          )}
-        </OutlinedBox>
-      </Box>
+      <DateBox />
 
       <Button variant='contained' color='primary' fullWidth>
         Vertretungen speichern
@@ -326,7 +201,7 @@ function TutorialSubstituteManagementContent({ tutorial, students }: Props): JSX
   );
 }
 
-function TutorialSubstituteManagement(): JSX.Element {
+function SubstituteManagement(): JSX.Element {
   const DUMMY_TUTORIAL_DATA: ITutorial = {
     id: 'dev-tutorial-id',
     slot: 'Mi07',
@@ -349,9 +224,7 @@ function TutorialSubstituteManagement(): JSX.Element {
   ];
   const DUMMY_TUTORIAL = plainToClass(Tutorial, DUMMY_TUTORIAL_DATA);
 
-  return (
-    <TutorialSubstituteManagementContent tutorial={DUMMY_TUTORIAL} students={DUMMY_STUDENTS} />
-  );
+  return <SubstituteManagementContent tutorial={DUMMY_TUTORIAL} students={DUMMY_STUDENTS} />;
 }
 
-export default TutorialSubstituteManagement;
+export default SubstituteManagement;
