+++ client/src/view/tutorial-substitutes/SubstituteManagement.types.ts
@@ -1,0 +1,6 @@
+export enum FilterOption {
+  ONLY_FUTURE_DATES = 'Nur zukünftige Termine',
+  ALL_DATES = 'Alle Termine',
+  WITHOUT_SUBSTITUTE = 'Ohne Vertretung',
+  WITH_SUBSTITUTE = 'Mit Vertretung',
+}
